requirements.txt
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
PyYAML==6.0.2
streamlit==1.39.0
